refactor: system metadata repository
