+++ server/src/services/backup.service.spec.ts
@@ -6,9 +6,8 @@ import { IDatabaseRepository } from 'src/interfaces/database.interface';
 import { JobStatus } from 'src/interfaces/job.interface';
 import { IProcessRepository } from 'src/interfaces/process.interface';
 import { IStorageRepository } from 'src/interfaces/storage.interface';
-import { ISystemMetadataRepository } from 'src/interfaces/system-metadata.interface';
 import { BackupService } from 'src/services/backup.service';
-import { IConfigRepository, ICronRepository } from 'src/types';
+import { IConfigRepository, ICronRepository, ISystemMetadataRepository } from 'src/types';
 import { systemConfigStub } from 'test/fixtures/system-config.stub';
 import { mockSpawn, newTestService } from 'test/utils';
 import { describe, Mocked } from 'vitest';
