+++ server/src/services/metadata.service.spec.ts
@@ -12,12 +12,17 @@ import { IEventRepository } from 'src/interfaces/event.interface';
 import { IJobRepository, JobName, JobStatus } from 'src/interfaces/job.interface';
 import { IPersonRepository } from 'src/interfaces/person.interface';
 import { IStorageRepository } from 'src/interfaces/storage.interface';
-import { ISystemMetadataRepository } from 'src/interfaces/system-metadata.interface';
 import { ITagRepository } from 'src/interfaces/tag.interface';
 import { IUserRepository } from 'src/interfaces/user.interface';
 import { ImmichTags } from 'src/repositories/metadata.repository';
 import { MetadataService } from 'src/services/metadata.service';
-import { IConfigRepository, IMapRepository, IMediaRepository, IMetadataRepository } from 'src/types';
+import {
+  IConfigRepository,
+  IMapRepository,
+  IMediaRepository,
+  IMetadataRepository,
+  ISystemMetadataRepository,
+} from 'src/types';
 import { assetStub } from 'test/fixtures/asset.stub';
 import { fileStub } from 'test/fixtures/file.stub';
 import { probeStub } from 'test/fixtures/media.stub';
