+++ server/src/services/storage.service.spec.ts
@@ -1,8 +1,7 @@
 import { SystemMetadataKey } from 'src/enum';
 import { IStorageRepository } from 'src/interfaces/storage.interface';
-import { ISystemMetadataRepository } from 'src/interfaces/system-metadata.interface';
 import { StorageService } from 'src/services/storage.service';
-import { IConfigRepository, ILoggingRepository } from 'src/types';
+import { IConfigRepository, ILoggingRepository, ISystemMetadataRepository } from 'src/types';
 import { ImmichStartupError } from 'src/utils/misc';
 import { mockEnvData } from 'test/repositories/config.repository.mock';
 import { newTestService } from 'test/utils';
