+++ server/src/services/version.service.spec.ts
@@ -4,9 +4,14 @@ import { serverVersion } from 'src/constants';
 import { ImmichEnvironment, SystemMetadataKey } from 'src/enum';
 import { IEventRepository } from 'src/interfaces/event.interface';
 import { IJobRepository, JobName, JobStatus } from 'src/interfaces/job.interface';
-import { ISystemMetadataRepository } from 'src/interfaces/system-metadata.interface';
 import { VersionService } from 'src/services/version.service';
-import { IConfigRepository, ILoggingRepository, IServerInfoRepository, IVersionHistoryRepository } from 'src/types';
+import {
+  IConfigRepository,
+  ILoggingRepository,
+  IServerInfoRepository,
+  ISystemMetadataRepository,
+  IVersionHistoryRepository,
+} from 'src/types';
 import { mockEnvData } from 'test/repositories/config.repository.mock';
 import { newTestService } from 'test/utils';
 import { Mocked } from 'vitest';
