+++ server/src/types.ts
@@ -15,6 +15,7 @@ import { MetadataRepository } from 'src/repositories/metadata.repository';
 import { NotificationRepository } from 'src/repositories/notification.repository';
 import { OAuthRepository } from 'src/repositories/oauth.repository';
 import { ServerInfoRepository } from 'src/repositories/server-info.repository';
+import { SystemMetadataRepository } from 'src/repositories/system-metadata.repository';
 import { MetricGroupRepository, TelemetryRepository } from 'src/repositories/telemetry.repository';
 import { TrashRepository } from 'src/repositories/trash.repository';
 import { VersionHistoryRepository } from 'src/repositories/version-history.repository';
@@ -59,6 +60,7 @@ export type IMetricGroupRepository = RepositoryInterface<MetricGroupRepository>;
 export type INotificationRepository = RepositoryInterface<NotificationRepository>;
 export type IOAuthRepository = RepositoryInterface<OAuthRepository>;
 export type IServerInfoRepository = RepositoryInterface<ServerInfoRepository>;
+export type ISystemMetadataRepository = RepositoryInterface<SystemMetadataRepository>;
 export type ITelemetryRepository = RepositoryInterface<TelemetryRepository>;
 export type ITrashRepository = RepositoryInterface<TrashRepository>;
 export type IViewRepository = RepositoryInterface<ViewRepository>;
